# api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Tenant PIM API
  version: 2.4.0
  description: |
    API para gestión de información de productos multi-tenant.

    ## Arquitectura Modular

    El sistema está organizado en módulos especializados:
    - **Health**: Estado del sistema
    - **Business Types**: Tipos de negocio
    - **Brands**: Gestión de marcas
    - **Products**: Gestión de productos y variantes
    - **Variants**: Gestión de variantes de productos (HITO 2.1)
    - **Categories**: Gestión de categorías
    - **Attributes**: Gestión de atributos
    - **Category-Attributes**: Relación categoría-atributo
    - **Global Catalog**: Catálogo global de productos con enrichment y quality score
    - **Global Catalog Public**: Endpoints públicos del catálogo global (sin auth)
    - **Product Requests**: Solicitudes de productos por tenants
    - **Quickstart**: Configuración rápida e importación desde global catalog
    - **Import**: Importación masiva con variantes (HITO 2.1)

    ## Actualizado v2.3.0

    - Global Catalog: CRUD completo, búsqueda por EAN, enrichment queue
    - Product Requests: Solicitudes de tenants con flujo approve/reject/fulfill
    - Quickstart: Import desde global catalog, import masivo por business type
    - Public endpoints: Búsqueda, sugerencias, facets sin autenticación
    - Variantes de productos: Product -> Variants con SKU único
    - Templates Quickstart: Configuración rápida por tipo de negocio
...
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para autenticación
    tenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: ID del tenant para operaciones multi-tenant
    userRole:
      type: apiKey
      in: header
      name: X-User-Role
      description: |
        Rol del usuario para autorización:
        - `super_admin`: Acceso completo
        - `marketplace_admin`: Gestión de marketplace
        - `tenant_admin`: Administración del tenant
        - `tenant_user`: Usuario del tenant
...
paths:
...
  /products/{id}/variants:
    get:
      summary: Listar variantes de un producto
      tags: [Variants]
      description: |
        Obtiene todas las variantes de un producto.
        Las variantes tienen SKU único y atributos dinámicos.
      security:
        - BearerAuth: []
        - tenantId: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: ID del producto padre
      responses:
        '200':
          description: Lista de variantes
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        product_id:
                          type: string
                          format: uuid
                        sku:
                          type: string
                          example: "TALADRO-BOSCH-750W"
                        name:
                          type: string
                          example: "Taladro Bosch 750W"
                        attributes:
                          type: object
                          additionalProperties: true
                          example:
                            potencia: "750W"
                            voltaje: "220V"
                        price:
                          type: number
                          example: 45000
                        status:
                          type: string
                          enum: ["active", "inactive", "discontinued"]
                        created_at:
                          type: string
                          format: date-time
                  total_count:
                    type: integer
        '404':
          description: Producto no encontrado
        '500':
          description: Error interno del servidor

    post:
      summary: Crear variante de producto
      tags: [Variants]
      description: Crea una nueva variante para un producto
      security:
        - BearerAuth: []
        - tenantId: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sku
                - name
              properties:
                sku:
                  type: string
                  example: "TALADRO-BOSCH-1000W"
                name:
                  type: string
                  example: "Taladro Bosch 1000W"
                attributes:
                  type: object
                  additionalProperties: true
                  example:
                    potencia: "1000W"
                    voltaje: "220V"
                price:
                  type: number
                  example: 65000
      responses:
        '201':
          description: Variante creada
        '400':
          description: Datos inválidos
        '404':
          description: Producto no encontrado
        '409':
          description: SKU ya existe
        '500':
          description: Error interno del servidor
  /variants/by-skus:
    post:
      summary: Búsqueda por lotes de variantes por SKU
      tags: [Variants]
      operationId: getVariantsBySKUs
      description: |
        Realiza una búsqueda por lotes de variantes de productos enriquecidas
        con información del producto y categoría.
      security:
        - BearerAuth: []
        - tenantId: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [skus]
              properties:
                skus:
                  type: array
                  items:
                    type: string
                  description: Lista de SKUs a buscar
                  example: ["SKU-1", "SKU-2", "SKU-3"]
      responses:
        '200':
          description: Lista de variantes enriquecidas encontradas
          content:
            application/json:
              schema:
                type: object
                required: [variants]
                properties:
                  variants:
                    type: array
                    items:
                      type: object
                      properties:
                        variant_id:
                          type: string
                          format: uuid
                        product_id:
                          type: string
                          format: uuid
                        sku:
                          type: string
                        variant_name:
                          type: string
                        product_name:
                          type: string
                        category_id:
                          type: string
                        category_name:
                          type: string
                        price:
                          type: number
                          format: float
        '400':
          description: Datos de entrada inválidos
        '401':
          description: No autorizado
        '500':
          description: Error interno del servidor
...